requirements.txt
streamlit>=1.29.0
alembic>=1.13.0
pydantic>=2.6.0
pandas>=2.2.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
plotly>=5.18.0
openpyxl>=3.1.0
